+++ srcs/parse/rt/parse_rt.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse_rt.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mfukui <mfukui@student.42.fr>              +#+  +:+       +#+        */
+/*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/15 22:15:03 by mfukui            #+#    #+#             */
-/*   Updated: 2025/04/22 16:42:09 by mfukui           ###   ########.fr       */
+/*   Updated: 2025/04/22 17:11:49 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,7 +70,7 @@ bool	parse_rt(t_rt *rt, char *file_name)
 		return (error_message(OBJECT), false);
 	rt->screen.width = SCREEN_WIDTH;
 	rt->screen.height = SCREEN_HEIGHT;
-	rt->screen.aspect_ratio = (float)SCREEN_WIDTH / (float)SCREEN_HEIGHT;
+	rt->screen.aspect_ratio = (double)SCREEN_WIDTH / (double)SCREEN_HEIGHT;
 	rt->screen.scale
 		= 1.0f / (2.0f * tanf(rt->camera->fov / 2.0f * (M_PI / 180.0f)));
 	return (true);
